Correcoes agendamento portador POS-PAGO e agendamento estab

+++ grails-app/domain/com/sysdata/gestaofrota/PostoCombustivel.groovy
@@ -2,7 +2,7 @@ package com.sysdata.gestaofrota
 
 class PostoCombustivel extends Empresa {
 	
-	Double taxaReembolso=0.0
+	BigDecimal taxaReembolso=0.0
 	
 	TipoReembolso tipoReembolso
